consume isLoading and success states and pass them as props to DeleteModal, so we i can then render jsx conditionally based on those states
Tried to consume the useDeleteModal hook directly in DeleteModal, but states were not updating. debugged the issue for a long time and finally realised that those states need to be passed as props - possibly the modal was not re-rendering and not listening for state changes. need to debug further to understand the issue in full. leaving it as it is for now as it works. refactor will come later on.