wrap the custom select in a container for easier manipulation of mainly css
also found a bug whereby if the custom select is placed on in column 2 (grid-column: 2 / 3) then is was overlapping with the toaster and i was wondering why its cursor: pointer was not working. in works in column 1 and 3, but overlaps with toaster in column 2, so for now i'll keep in in column one.  it will eventually be completely removed from there, so it does not matter, but still.