fix a bug that was causing multiple intervals to be started
this resulted in issues with body scrolling after success modal disappeared; it happened after successful login, register, edit or adding a new quote

I wondered why this was and it turned our that i had extremely dumb conditional check:

if (seconds === delayMs / 1000) {
 startTimer();
}

Issue with this was that this conditional check is ALWAYS true, which caused the inconsistent behavior of the body scroll, but more importantly the app was spawning multiple intervals (depending on count number) and then clearing only one.

replaced this by checking if we already have an intervalId. If not - only then call startTimer fn;